Move types around to fix import loops, add logging
Replace logging in both binaries with LoggerArguments
